remove 'bootstrap' as an input
Our self-hosted runner needed this by its design; these builds
will execute within independent windows containers so we need
to always run the cerbero bootstrap command since it's part of
the initial configuration of the environment prior to the build.
Also changed the id of the build step to something clearer.